@@ -22,7 +22,7 @@ namespace CodeGenerator
             }
             else
             {
-                outputPath = Path.Combine(AppContext.BaseDirectory, "generated");
+                outputPath = AppContext.BaseDirectory;
             }
 
             if (!Directory.Exists(outputPath))
@@ -75,7 +75,7 @@ namespace CodeGenerator
                 "cimguizmo" => "cimguizmo",
                 _ => throw new NotImplementedException()
             };
-
+            
             string definitionsPath = Path.Combine(AppContext.BaseDirectory, "definitions", libraryName);
             var defs = new ImguiDefinitions();
             defs.LoadFrom(definitionsPath);
@@ -84,14 +84,14 @@ namespace CodeGenerator
 
             foreach (EnumDefinition ed in defs.Enums)
             {
-                using (CSharpCodeWriter writer = new CSharpCodeWriter(Path.Combine(outputPath, ed.FriendlyName + ".gen.cs")))
+                using (CSharpCodeWriter writer = new CSharpCodeWriter(Path.Combine(outputPath, ed.FriendlyNames[0] + ".gen.cs")))
                 {
                     writer.PushBlock($"namespace {projectNamespace}");
-                    if (ed.FriendlyName.Contains("Flags"))
+                    if (ed.FriendlyNames[0].Contains("Flags"))
                     {
                         writer.WriteLine("[System.Flags]");
                     }
-                    writer.PushBlock($"public enum {ed.FriendlyName}");
+                    writer.PushBlock($"public enum {ed.FriendlyNames[0]}");
                     foreach (EnumMember member in ed.Members)
                     {
                         string sanitizedName = ed.SanitizeNames(member.Name);
@@ -373,6 +373,12 @@ namespace CodeGenerator
                         }
                         if (exportedName.Contains("~")) { continue; }
                         if (overload.Parameters.Any(tr => tr.Type.Contains('('))) { continue; } // TODO: Parse function pointer parameters.
+                        
+                        if ((overload.FriendlyName == "GetID" || overload.FriendlyName == "PushID") && overload.Parameters.Length > 1)
+                        {
+                            // skip ImGui.Get/PushID(start, end) overloads as they would overlap with existing
+                            continue;
+                        }
 
                         bool hasVaList = false;
                         for (int i = 0; i < overload.Parameters.Length; i++)
@@ -448,9 +454,22 @@ namespace CodeGenerator
             string selfName,
             string classPrefix)
         {
-            if (overload.Parameters.Where(tr => tr.Name.EndsWith("_begin") || tr.Name.EndsWith("_end"))
-                .Any(tr => !defaultValues.ContainsKey(tr.Name)))
+            var rangeParams = overload.Parameters.Where(tr => 
+                tr.Name.EndsWith("_begin") || 
+                tr.Name.EndsWith("_end")).ToArray();
+            if (rangeParams.Any(tr => tr.Type != "char*"))
             {
+                // only string supported for start/end. ImFont.IsGlyphRangeUnused is uint
+                return; 
+            }
+            if (rangeParams.Any(tr => tr.Name.EndsWith("_end") && defaultValues.ContainsKey(tr.Name)))
+            {
+                // we want overloads:
+                // (something, text_start, text_end (deleted), after=default)
+                // (something, text_start, text_end (deleted), after)
+                
+                // we dont need (as it would be duplicate)
+                // (something, text_start, text_end=default (deleted))
                 return;
             }
 
@@ -463,7 +482,7 @@ namespace CodeGenerator
                 safeRet = GetSafeType(overload.ReturnType);
             }
 
-            List<string> invocationArgs = new List<string>();
+            List<(string MarshalledType, string CorrectedIdentifier)> invocationArgs = new();
             MarshalledParameter[] marshalledParameters = new MarshalledParameter[overload.Parameters.Length];
             List<string> preCallLines = new List<string>();
             List<string> postCallLines = new List<string>();
@@ -477,7 +496,7 @@ namespace CodeGenerator
                 if (tr.Name == "self")
                 {
                     selfIndex = i;
-                    continue;
+                    continue; 
                 }
                 if (tr.Name == "...") { continue; }
 
@@ -514,9 +533,20 @@ namespace CodeGenerator
                     }
                     else
                     {
+                        if (tr.Name.EndsWith("_end"))
+                        {
+                            var startParamName = overload.Parameters[i-1].Name;
+                            var startNativeParamName = $"native_{startParamName}";
+                            marshalledParameters[i] = new MarshalledParameter(nativeTypeName, false, $"{startNativeParamName}+{startParamName}_byteCount", false);
+                            continue;
+                        }
+                        
+                        var checkForNull = !hasDefault && !tr.Name.EndsWith("_begin");
+                        // for string _begin the pointer passed must be non-null, so we'll set up an empty string if needed
+                        
                         preCallLines.Add($"byte* {nativeArgName};");
                         preCallLines.Add($"int {correctedIdentifier}_byteCount = 0;");
-                        if (!hasDefault)
+                        if (checkForNull)
                         {
                             preCallLines.Add($"if ({textToEncode} != null)");
                             preCallLines.Add("{");
@@ -534,7 +564,7 @@ namespace CodeGenerator
                         preCallLines.Add($"    int {nativeArgName}_offset = Util.GetUtf8({textToEncode}, {nativeArgName}, {correctedIdentifier}_byteCount);");
                         preCallLines.Add($"    {nativeArgName}[{nativeArgName}_offset] = 0;");
 
-                        if (!hasDefault)
+                        if (checkForNull)
                         {
                             preCallLines.Add("}");
                             preCallLines.Add($"else {{ {nativeArgName} = null; }}");
@@ -576,12 +606,8 @@ namespace CodeGenerator
                     preCallLines.Add($"for (int i = 0; i < {correctedIdentifier}.Length; i++)");
                     preCallLines.Add("{");
                     preCallLines.Add($"    string s = {correctedIdentifier}[i];");
-                    preCallLines.Add($"    fixed (char* sPtr = s)");
-                    preCallLines.Add("    {");
-                    preCallLines.Add($"        offset += Encoding.UTF8.GetBytes(sPtr, s.Length, {nativeArgName}_data + offset, {correctedIdentifier}_byteCounts[i]);");
-                    preCallLines.Add($"        {nativeArgName}_data[offset] = 0;");
-                    preCallLines.Add($"        offset += 1;");
-                    preCallLines.Add("    }");
+                    preCallLines.Add($"    offset += Util.GetUtf8(s, {nativeArgName}_data + offset, {correctedIdentifier}_byteCounts[i]);");
+                    preCallLines.Add($"    {nativeArgName}_data[offset++] = 0;");
                     preCallLines.Add("}");
 
                     preCallLines.Add($"byte** {nativeArgName} = stackalloc byte*[{correctedIdentifier}.Length];");
@@ -622,7 +648,7 @@ namespace CodeGenerator
                     marshalledParameters[i] = new MarshalledParameter(wrappedParamType, false, nativeArgName, false);
                     preCallLines.Add($"{tr.Type} {nativeArgName} = {correctedIdentifier}.NativePtr;");
                 }
-                else if ((tr.Type.EndsWith("*") || tr.Type.Contains("[") || tr.Type.EndsWith("&")) && tr.Type != "void*" && tr.Type != "ImGuiContext*" && tr.Type != "ImPlotContext*" && tr.Type != "EditorContext*")
+                else if ((tr.Type.EndsWith("*") || tr.Type.Contains("[") || tr.Type.EndsWith("&")) && tr.Type != "void*" && tr.Type != "ImGuiContext*" && tr.Type != "ImPlotContext*"&& tr.Type != "EditorContext*")
                 {
                     string nonPtrType;
                     if (tr.Type.Contains("["))
@@ -647,21 +673,54 @@ namespace CodeGenerator
 
                 if (!marshalledParameters[i].HasDefaultValue)
                 {
-                    invocationArgs.Add($"{marshalledParameters[i].MarshalledType} {correctedIdentifier}");
+                    invocationArgs.Add((marshalledParameters[i].MarshalledType, correctedIdentifier));
                 }
             }
 
-            string invocationList = string.Join(", ", invocationArgs);
+            string invocationList = string.Join(", ", invocationArgs.Select(a => $"{a.MarshalledType} {a.CorrectedIdentifier}"));
             string friendlyName = overload.FriendlyName;
 
             string staticPortion = selfName == null ? "static " : string.Empty;
+            
+            // When .NET Standard 2.1 is available, we can use ReadOnlySpan<char> instead of string, so generate additional overloads for methods containing string parameters.
+            if (invocationArgs.Count > 0 && invocationArgs.Any(a => a is { MarshalledType: "string" }))
+            {
+                string readOnlySpanInvocationList = string.Join(", ", invocationArgs.Select(a => $"{(a.MarshalledType == "string" ? "ReadOnlySpan<char>" : a.MarshalledType)} {a.CorrectedIdentifier}"));
+
+                writer.WriteRaw("#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP2_1_OR_GREATER");
+                WriteMethod(writer, overload, selfName, classPrefix, staticPortion, overrideRet, safeRet, friendlyName, readOnlySpanInvocationList, preCallLines, marshalledParameters, selfIndex, pOutIndex, nativeRet, postCallLines, isWrappedType);
+                writer.WriteRaw("#endif");
+            }
+
+            WriteMethod(writer, overload, selfName, classPrefix, staticPortion, overrideRet, safeRet, friendlyName, invocationList, preCallLines, marshalledParameters, selfIndex, pOutIndex, nativeRet, postCallLines, isWrappedType);
+        }
+
+        private static void WriteMethod(
+            CSharpCodeWriter writer,
+            OverloadDefinition overload,
+            string selfName,
+            string classPrefix,
+            string staticPortion,
+            string overrideRet,
+            string safeRet,
+            string friendlyName,
+            string invocationList,
+            List<string> preCallLines,
+            MarshalledParameter[] marshalledParameters,
+            int selfIndex,
+            int pOutIndex,
+            string nativeRet,
+            List<string> postCallLines,
+            bool isWrappedType)
+        {
             writer.PushBlock($"public {staticPortion}{overrideRet ?? safeRet} {friendlyName}({invocationList})");
+
             foreach (string line in preCallLines)
             {
                 writer.WriteLine(line);
             }
 
-            List<string> nativeInvocationArgs = new List<string>();
+            List<string> nativeInvocationArgs = new();
 
             for (int i = 0; i < marshalledParameters.Length; i++)
             {
