@@ -56,13 +56,24 @@ namespace CodeGenerator
             { "ImPlotGetter", "IntPtr" },
             { "ImPlotTransform", "IntPtr" },
             { "ImGuiKeyChord", "ImGuiKey" },
+            { "ImGuiSelectionUserData", "long" },
         };
         
         public static readonly List<string> WellKnownEnums = new List<string>()
         {
             "ImGuiMouseButton"
         };
-        
+
+        public static readonly Dictionary<string, string> AlternateEnumPrefixes = new Dictionary<string, string>()
+        {
+            { "ImGuiKey", "ImGuiMod" },
+        };
+
+        public static readonly Dictionary<string, string> AlternateEnumPrefixSubstitutions = new Dictionary<string, string>()
+        {
+            { "ImGuiMod_", "Mod" },
+        };
+
         public static readonly Dictionary<string, string> WellKnownFieldReplacements = new Dictionary<string, string>()
         {
             { "bool", "bool" }, // Force bool to remain as bool in type-safe wrappers.
@@ -148,7 +159,6 @@ namespace CodeGenerator
         {
             "igInputText",
             "igInputTextMultiline",
-            "igCalcTextSize",
             "igInputTextWithHint"
         };
     }
