@@ -5,11 +5,6 @@ using System.Reflection;
 using System.IO;
 using Veldrid;
 using System.Runtime.CompilerServices;
-using ImGui.NET.SampleProgram;
-using Veldrid.Sdl2;
-using System.Runtime.InteropServices;
-using ImGuiNET;
-using ImPlotNET;
 
 namespace ImGuiNET
 {
@@ -20,7 +15,6 @@ namespace ImGuiNET
     public class ImGuiController : IDisposable
     {
         private GraphicsDevice _gd;
-        private readonly Sdl2Window _window;
         private bool _frameBegun;
 
         // Veldrid objects
@@ -54,190 +48,28 @@ namespace ImGuiNET
             = new Dictionary<Texture, TextureView>();
         private readonly Dictionary<IntPtr, ResourceSetInfo> _viewsById = new Dictionary<IntPtr, ResourceSetInfo>();
         private readonly List<IDisposable> _ownedResources = new List<IDisposable>();
-        private readonly VeldridImGuiWindow _mainViewportWindow;
-        private readonly Platform_CreateWindow _createWindow;
-        private readonly Platform_DestroyWindow _destroyWindow;
-        private readonly Platform_GetWindowPos _getWindowPos;
-        private readonly Platform_ShowWindow _showWindow;
-        private readonly Platform_SetWindowPos _setWindowPos;
-        private readonly Platform_SetWindowSize _setWindowSize;
-        private readonly Platform_GetWindowSize _getWindowSize;
-        private readonly Platform_SetWindowFocus _setWindowFocus;
-        private readonly Platform_GetWindowFocus _getWindowFocus;
-        private readonly Platform_GetWindowMinimized _getWindowMinimized;
-        private readonly Platform_SetWindowTitle _setWindowTitle;
         private int _lastAssignedID = 100;
 
         /// <summary>
         /// Constructs a new ImGuiController.
         /// </summary>
-        public unsafe ImGuiController(GraphicsDevice gd, Sdl2Window window, OutputDescription outputDescription, int width, int height)
+        public ImGuiController(GraphicsDevice gd, OutputDescription outputDescription, int width, int height)
         {
             _gd = gd;
-            _window = window;
             _windowWidth = width;
             _windowHeight = height;
 
-            //Init imgui context
-            IntPtr context = ImGui.CreateContext();
-            ImGui.SetCurrentContext(context);
-
-            //Init implot context
-            IntPtr plotcontext = ImPlot.CreateContext();
-            ImPlot.SetImGuiContext(context);
-            ImGuiIOPtr io = ImGui.GetIO();
-
-            io.ConfigFlags |= ImGuiConfigFlags.DockingEnable;
-            io.ConfigFlags |= ImGuiConfigFlags.ViewportsEnable;
-
-            ImGuiPlatformIOPtr platformIO = ImGui.GetPlatformIO();
-            ImGuiViewportPtr mainViewport = platformIO.Viewports[0];
-            mainViewport.PlatformHandle = window.Handle;
-            _mainViewportWindow = new VeldridImGuiWindow(gd, mainViewport, _window);
-
-            _createWindow = CreateWindow;
-            _destroyWindow = DestroyWindow;
-            _getWindowPos = GetWindowPos;
-            _showWindow = ShowWindow;
-            _setWindowPos = SetWindowPos;
-            _setWindowSize = SetWindowSize;
-            _getWindowSize = GetWindowSize;
-            _setWindowFocus = SetWindowFocus;
-            _getWindowFocus = GetWindowFocus;
-            _getWindowMinimized = GetWindowMinimized;
-            _setWindowTitle = SetWindowTitle;
-
-            platformIO.Platform_CreateWindow = Marshal.GetFunctionPointerForDelegate(_createWindow);
-            platformIO.Platform_DestroyWindow = Marshal.GetFunctionPointerForDelegate(_destroyWindow);
-            platformIO.Platform_ShowWindow = Marshal.GetFunctionPointerForDelegate(_showWindow);
-            platformIO.Platform_SetWindowPos = Marshal.GetFunctionPointerForDelegate(_setWindowPos);
-            platformIO.Platform_SetWindowSize = Marshal.GetFunctionPointerForDelegate(_setWindowSize);
-            platformIO.Platform_SetWindowFocus = Marshal.GetFunctionPointerForDelegate(_setWindowFocus);
-            platformIO.Platform_GetWindowFocus = Marshal.GetFunctionPointerForDelegate(_getWindowFocus);
-            platformIO.Platform_GetWindowMinimized = Marshal.GetFunctionPointerForDelegate(_getWindowMinimized);
-            platformIO.Platform_SetWindowTitle = Marshal.GetFunctionPointerForDelegate(_setWindowTitle);
-
-            ImGuiNative.ImGuiPlatformIO_Set_Platform_GetWindowPos(platformIO.NativePtr, Marshal.GetFunctionPointerForDelegate(_getWindowPos));
-            ImGuiNative.ImGuiPlatformIO_Set_Platform_GetWindowSize(platformIO.NativePtr, Marshal.GetFunctionPointerForDelegate(_getWindowSize));
-
-            unsafe
-            {
-                io.NativePtr->BackendPlatformName = (byte*)new FixedAsciiString("Veldrid.SDL2 Backend").DataPtr;
-            }
-            io.BackendFlags |= ImGuiBackendFlags.HasMouseCursors;
-            io.BackendFlags |= ImGuiBackendFlags.HasSetMousePos;
-            io.BackendFlags |= ImGuiBackendFlags.PlatformHasViewports;
-            io.BackendFlags |= ImGuiBackendFlags.RendererHasViewports;
+            ImGui.CreateContext();
+            var io = ImGui.GetIO();
             io.BackendFlags |= ImGuiBackendFlags.RendererHasVtxOffset;
-
-            var fonts = io.Fonts;
-            io.Fonts.AddFontDefault();
+            io.ConfigFlags |= ImGuiConfigFlags.NavEnableKeyboard |
+                ImGuiConfigFlags.DockingEnable;
+            io.Fonts.Flags |= ImFontAtlasFlags.NoBakedLines;
 
             CreateDeviceResources(gd, outputDescription);
-            SetKeyMappings();
-
             SetPerFrameImGuiData(1f / 60f);
-            UpdateMonitors();
-
             ImGui.NewFrame();
             _frameBegun = true;
-        }
-
-        private void CreateWindow(ImGuiViewportPtr vp)
-        {
-            VeldridImGuiWindow window = new VeldridImGuiWindow(_gd, vp);
-        }
-
-        private void DestroyWindow(ImGuiViewportPtr vp)
-        {
-            if (vp.PlatformUserData != IntPtr.Zero)
-            {
-                VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-                window.Dispose();
-
-                vp.PlatformUserData = IntPtr.Zero;
-            }
-        }
-
-        private void ShowWindow(ImGuiViewportPtr vp)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            Sdl2Native.SDL_ShowWindow(window.Window.SdlWindowHandle);
-        }
-
-        private unsafe void GetWindowPos(ImGuiViewportPtr vp, Vector2* outPos)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            *outPos = new Vector2(window.Window.Bounds.X, window.Window.Bounds.Y);
-        }
-
-        private void SetWindowPos(ImGuiViewportPtr vp, Vector2 pos)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            window.Window.X = (int)pos.X;
-            window.Window.Y = (int)pos.Y;
-        }
-
-        private void SetWindowSize(ImGuiViewportPtr vp, Vector2 size)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            Sdl2Native.SDL_SetWindowSize(window.Window.SdlWindowHandle, (int)size.X, (int)size.Y);
-        }
-
-        private unsafe void GetWindowSize(ImGuiViewportPtr vp, Vector2* outSize)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            Rectangle bounds = window.Window.Bounds;
-            *outSize = new Vector2(bounds.Width, bounds.Height);
-        }
-
-        private delegate void SDL_RaiseWindow_t(IntPtr sdl2Window);
-        private static SDL_RaiseWindow_t p_sdl_RaiseWindow;
-
-        private unsafe delegate uint SDL_GetGlobalMouseState_t(int* x, int* y);
-        private static SDL_GetGlobalMouseState_t p_sdl_GetGlobalMouseState;
-
-        private unsafe delegate int SDL_GetDisplayUsableBounds_t(int displayIndex, Rectangle* rect);
-        private static SDL_GetDisplayUsableBounds_t p_sdl_GetDisplayUsableBounds_t;
-
-        private delegate int SDL_GetNumVideoDisplays_t();
-        private static SDL_GetNumVideoDisplays_t p_sdl_GetNumVideoDisplays;
-
-        private void SetWindowFocus(ImGuiViewportPtr vp)
-        {
-            if (p_sdl_RaiseWindow == null)
-            {
-                p_sdl_RaiseWindow = Sdl2Native.LoadFunction<SDL_RaiseWindow_t>("SDL_RaiseWindow");
-            }
-
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            p_sdl_RaiseWindow(window.Window.SdlWindowHandle);
-        }
-
-        private byte GetWindowFocus(ImGuiViewportPtr vp)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            SDL_WindowFlags flags = Sdl2Native.SDL_GetWindowFlags(window.Window.SdlWindowHandle);
-            return (flags & SDL_WindowFlags.InputFocus) != 0 ? (byte)1 : (byte)0;
-        }
-
-        private byte GetWindowMinimized(ImGuiViewportPtr vp)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            SDL_WindowFlags flags = Sdl2Native.SDL_GetWindowFlags(window.Window.SdlWindowHandle);
-            return (flags & SDL_WindowFlags.Minimized) != 0 ? (byte)1 : (byte)0;
-        }
-
-        private unsafe void SetWindowTitle(ImGuiViewportPtr vp, IntPtr title)
-        {
-            VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-            byte* titlePtr = (byte*)title;
-            int count = 0;
-            while (titlePtr[count] != 0)
-            {
-                count += 1;
-            }
-            window.Window.Title = System.Text.Encoding.ASCII.GetString(titlePtr, count);
         }
 
         public void WindowResized(int width, int height)
@@ -412,11 +244,11 @@ namespace ImGuiNET
         /// <summary>
         /// Recreates the device texture used to render text.
         /// </summary>
-        public unsafe void RecreateFontDeviceTexture(GraphicsDevice gd)
+        public void RecreateFontDeviceTexture(GraphicsDevice gd)
         {
             ImGuiIOPtr io = ImGui.GetIO();
             // Build
-            byte* pixels;
+            IntPtr pixels;
             int width, height, bytesPerPixel;
             io.Fonts.GetTexDataAsRGBA32(out pixels, out width, out height, out bytesPerPixel);
             // Store our identifier
@@ -432,7 +264,7 @@ namespace ImGuiNET
             _fontTexture.Name = "ImGui.NET Font Texture";
             gd.UpdateTexture(
                 _fontTexture,
-                (IntPtr)pixels,
+                pixels,
                 (uint)(bytesPerPixel * width * height),
                 0,
                 0,
@@ -460,31 +292,6 @@ namespace ImGuiNET
                 _frameBegun = false;
                 ImGui.Render();
                 RenderImDrawData(ImGui.GetDrawData(), gd, cl);
-
-                // Update and Render additional Platform Windows
-                if ((ImGui.GetIO().ConfigFlags & ImGuiConfigFlags.ViewportsEnable) != 0)
-                {
-                    ImGui.UpdatePlatformWindows();
-                    ImGuiPlatformIOPtr platformIO = ImGui.GetPlatformIO();
-                    for (int i = 1; i < platformIO.Viewports.Size; i++)
-                    {
-                        ImGuiViewportPtr vp = platformIO.Viewports[i];
-                        VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-                        cl.SetFramebuffer(window.Swapchain.Framebuffer);
-                        RenderImDrawData(vp.DrawData, gd, cl);
-                    }
-                }
-            }
-        }
-
-        public void SwapExtraWindows(GraphicsDevice gd)
-        {
-            ImGuiPlatformIOPtr platformIO = ImGui.GetPlatformIO();
-            for (int i = 1; i < platformIO.Viewports.Size; i++)
-            {
-                ImGuiViewportPtr vp = platformIO.Viewports[i];
-                VeldridImGuiWindow window = (VeldridImGuiWindow)GCHandle.FromIntPtr(vp.PlatformUserData).Target;
-                gd.SwapBuffers(window.Swapchain);
             }
         }
 
@@ -496,48 +303,13 @@ namespace ImGuiNET
             if (_frameBegun)
             {
                 ImGui.Render();
-                ImGui.UpdatePlatformWindows();
             }
 
             SetPerFrameImGuiData(deltaSeconds);
             UpdateImGuiInput(snapshot);
-            UpdateMonitors();
 
             _frameBegun = true;
             ImGui.NewFrame();
-
-            ImGui.Text($"Main viewport Position: {ImGui.GetPlatformIO().Viewports[0].Pos}");
-            ImGui.Text($"Main viewport Size: {ImGui.GetPlatformIO().Viewports[0].Size}");
-            ImGui.Text($"MouseHoveredViewport: {ImGui.GetIO().MouseHoveredViewport}");
-        }
-
-        private unsafe void UpdateMonitors()
-        {
-            if (p_sdl_GetNumVideoDisplays == null)
-            {
-                p_sdl_GetNumVideoDisplays = Sdl2Native.LoadFunction<SDL_GetNumVideoDisplays_t>("SDL_GetNumVideoDisplays");
-            }
-            if (p_sdl_GetDisplayUsableBounds_t == null)
-            {
-                p_sdl_GetDisplayUsableBounds_t = Sdl2Native.LoadFunction<SDL_GetDisplayUsableBounds_t>("SDL_GetDisplayUsableBounds");
-            }
-
-            ImGuiPlatformIOPtr platformIO = ImGui.GetPlatformIO();
-            Marshal.FreeHGlobal(platformIO.NativePtr->Monitors.Data);
-            int numMonitors = p_sdl_GetNumVideoDisplays();
-            IntPtr data = Marshal.AllocHGlobal(Unsafe.SizeOf<ImGuiPlatformMonitor>() * numMonitors);
-            platformIO.NativePtr->Monitors = new ImVector(numMonitors, numMonitors, data);
-            for (int i = 0; i < numMonitors; i++)
-            {
-                Rectangle r;
-                p_sdl_GetDisplayUsableBounds_t(i, &r);
-                ImGuiPlatformMonitorPtr monitor = platformIO.Monitors[i];
-                monitor.DpiScale = 1f;
-                monitor.MainPos = new Vector2(r.X, r.Y);
-                monitor.MainSize = new Vector2(r.Width, r.Height);
-                monitor.WorkPos = new Vector2(r.X, r.Y);
-                monitor.WorkSize = new Vector2(r.Width, r.Height);
-            }
         }
 
         /// <summary>
@@ -552,129 +324,93 @@ namespace ImGuiNET
                 _windowHeight / _scaleFactor.Y);
             io.DisplayFramebufferScale = _scaleFactor;
             io.DeltaTime = deltaSeconds; // DeltaTime is in seconds.
+        }
 
-            ImGui.GetPlatformIO().Viewports[0].Pos = new Vector2(_window.X, _window.Y);
-            ImGui.GetPlatformIO().Viewports[0].Size = new Vector2(_window.Width, _window.Height);
+        private bool TryMapKey(Key key, out ImGuiKey result)
+        {
+            ImGuiKey KeyToImGuiKeyShortcut(Key keyToConvert, Key startKey1, ImGuiKey startKey2)
+            {
+                int changeFromStart1 = (int)keyToConvert - (int)startKey1;
+                return startKey2 + changeFromStart1;
+            }
+
+            result = key switch
+            {
+                >= Key.F1 and <= Key.F24 => KeyToImGuiKeyShortcut(key, Key.F1, ImGuiKey.F1),
+                >= Key.Keypad0 and <= Key.Keypad9 => KeyToImGuiKeyShortcut(key, Key.Keypad0, ImGuiKey.Keypad0),
+                >= Key.A and <= Key.Z => KeyToImGuiKeyShortcut(key, Key.A, ImGuiKey.A),
+                >= Key.Number0 and <= Key.Number9 => KeyToImGuiKeyShortcut(key, Key.Number0, ImGuiKey._0),
+                Key.ShiftLeft or Key.ShiftRight => ImGuiKey.ModShift,
+                Key.ControlLeft or Key.ControlRight => ImGuiKey.ModCtrl,
+                Key.AltLeft or Key.AltRight => ImGuiKey.ModAlt,
+                Key.WinLeft or Key.WinRight => ImGuiKey.ModSuper,
+                Key.Menu => ImGuiKey.Menu,
+                Key.Up => ImGuiKey.UpArrow,
+                Key.Down => ImGuiKey.DownArrow,
+                Key.Left => ImGuiKey.LeftArrow,
+                Key.Right => ImGuiKey.RightArrow,
+                Key.Enter => ImGuiKey.Enter,
+                Key.Escape => ImGuiKey.Escape,
+                Key.Space => ImGuiKey.Space,
+                Key.Tab => ImGuiKey.Tab,
+                Key.BackSpace => ImGuiKey.Backspace,
+                Key.Insert => ImGuiKey.Insert,
+                Key.Delete => ImGuiKey.Delete,
+                Key.PageUp => ImGuiKey.PageUp,
+                Key.PageDown => ImGuiKey.PageDown,
+                Key.Home => ImGuiKey.Home,
+                Key.End => ImGuiKey.End,
+                Key.CapsLock => ImGuiKey.CapsLock,
+                Key.ScrollLock => ImGuiKey.ScrollLock,
+                Key.PrintScreen => ImGuiKey.PrintScreen,
+                Key.Pause => ImGuiKey.Pause,
+                Key.NumLock => ImGuiKey.NumLock,
+                Key.KeypadDivide => ImGuiKey.KeypadDivide,
+                Key.KeypadMultiply => ImGuiKey.KeypadMultiply,
+                Key.KeypadSubtract => ImGuiKey.KeypadSubtract,
+                Key.KeypadAdd => ImGuiKey.KeypadAdd,
+                Key.KeypadDecimal => ImGuiKey.KeypadDecimal,
+                Key.KeypadEnter => ImGuiKey.KeypadEnter,
+                Key.Tilde => ImGuiKey.GraveAccent,
+                Key.Minus => ImGuiKey.Minus,
+                Key.Plus => ImGuiKey.Equal,
+                Key.BracketLeft => ImGuiKey.LeftBracket,
+                Key.BracketRight => ImGuiKey.RightBracket,
+                Key.Semicolon => ImGuiKey.Semicolon,
+                Key.Quote => ImGuiKey.Apostrophe,
+                Key.Comma => ImGuiKey.Comma,
+                Key.Period => ImGuiKey.Period,
+                Key.Slash => ImGuiKey.Slash,
+                Key.BackSlash or Key.NonUSBackSlash => ImGuiKey.Backslash,
+                _ => ImGuiKey.None
+            };
+
+            return result != ImGuiKey.None;
         }
 
         private void UpdateImGuiInput(InputSnapshot snapshot)
         {
             ImGuiIOPtr io = ImGui.GetIO();
-
-            Vector2 mousePosition = snapshot.MousePosition;
-
-            // Determine if any of the mouse buttons were pressed during this snapshot period, even if they are no longer held.
-            bool leftPressed = false;
-            bool middlePressed = false;
-            bool rightPressed = false;
-            foreach (MouseEvent me in snapshot.MouseEvents)
+            io.AddMousePosEvent(snapshot.MousePosition.X, snapshot.MousePosition.Y);
+            io.AddMouseButtonEvent(0, snapshot.IsMouseDown(MouseButton.Left));
+            io.AddMouseButtonEvent(1, snapshot.IsMouseDown(MouseButton.Right));
+            io.AddMouseButtonEvent(2, snapshot.IsMouseDown(MouseButton.Middle));
+            io.AddMouseButtonEvent(3, snapshot.IsMouseDown(MouseButton.Button1));
+            io.AddMouseButtonEvent(4, snapshot.IsMouseDown(MouseButton.Button2));
+            io.AddMouseWheelEvent(0f, snapshot.WheelDelta);
+            for (int i = 0; i < snapshot.KeyCharPresses.Count; i++)
             {
-                if (me.Down)
+                io.AddInputCharacter(snapshot.KeyCharPresses[i]);
+            }
+
+            for (int i = 0; i < snapshot.KeyEvents.Count; i++)
+            {
+                KeyEvent keyEvent = snapshot.KeyEvents[i];
+                if (TryMapKey(keyEvent.Key, out ImGuiKey imguikey))
                 {
-                    switch (me.MouseButton)
-                    {
-                        case MouseButton.Left:
-                            leftPressed = true;
-                            break;
-                        case MouseButton.Middle:
-                            middlePressed = true;
-                            break;
-                        case MouseButton.Right:
-                            rightPressed = true;
-                            break;
-                    }
+                    io.AddKeyEvent(imguikey, keyEvent.Down);
                 }
             }
-
-            io.MouseDown[0] = leftPressed || snapshot.IsMouseDown(MouseButton.Left);
-            io.MouseDown[1] = middlePressed || snapshot.IsMouseDown(MouseButton.Right);
-            io.MouseDown[2] = rightPressed || snapshot.IsMouseDown(MouseButton.Middle);
-
-            if (p_sdl_GetGlobalMouseState == null)
-            {
-                p_sdl_GetGlobalMouseState = Sdl2Native.LoadFunction<SDL_GetGlobalMouseState_t>("SDL_GetGlobalMouseState");
-            }
-
-            int x, y;
-            unsafe
-            {
-                uint buttons = p_sdl_GetGlobalMouseState(&x, &y);
-                io.MouseDown[0] = (buttons & 0b0001) != 0;
-                io.MouseDown[1] = (buttons & 0b0010) != 0;
-                io.MouseDown[2] = (buttons & 0b0100) != 0;
-            }
-
-            io.MousePos = new Vector2(x, y);
-            io.MouseWheel = snapshot.WheelDelta;
-            io = ImGui.GetIO();
-            IReadOnlyList<char> keyCharPresses = snapshot.KeyCharPresses;
-            for (int i = 0; i < keyCharPresses.Count; i++)
-            {
-                char c = keyCharPresses[i];
-                io.AddInputCharacter(c);
-            }
-
-            IReadOnlyList<KeyEvent> keyEvents = snapshot.KeyEvents;
-            for (int i = 0; i < keyEvents.Count; i++)
-            {
-                KeyEvent keyEvent = keyEvents[i];
-                io.KeysDown[(int)keyEvent.Key] = keyEvent.Down;
-                if (keyEvent.Key == Key.ControlLeft)
-                {
-                    _controlDown = keyEvent.Down;
-                }
-                if (keyEvent.Key == Key.ShiftLeft)
-                {
-                    _shiftDown = keyEvent.Down;
-                }
-                if (keyEvent.Key == Key.AltLeft)
-                {
-                    _altDown = keyEvent.Down;
-                }
-                if (keyEvent.Key == Key.WinLeft)
-                {
-                    _winKeyDown = keyEvent.Down;
-                }
-            }
-
-            io.KeyCtrl = _controlDown;
-            io.KeyAlt = _altDown;
-            io.KeyShift = _shiftDown;
-            io.KeySuper = _winKeyDown;
-
-            ImVector<ImGuiViewportPtr> viewports = ImGui.GetPlatformIO().Viewports;
-            for (int i = 1; i < viewports.Size; i++)
-            {
-                ImGuiViewportPtr v = viewports[i];
-                VeldridImGuiWindow window = ((VeldridImGuiWindow)GCHandle.FromIntPtr(v.PlatformUserData).Target);
-                window.Update();
-            }
-        }
-
-        private static void SetKeyMappings()
-        {
-            ImGuiIOPtr io = ImGui.GetIO();
-            io.KeyMap[(int)ImGuiKey.Tab] = (int)Key.Tab;
-            io.KeyMap[(int)ImGuiKey.LeftArrow] = (int)Key.Left;
-            io.KeyMap[(int)ImGuiKey.RightArrow] = (int)Key.Right;
-            io.KeyMap[(int)ImGuiKey.UpArrow] = (int)Key.Up;
-            io.KeyMap[(int)ImGuiKey.DownArrow] = (int)Key.Down;
-            io.KeyMap[(int)ImGuiKey.PageUp] = (int)Key.PageUp;
-            io.KeyMap[(int)ImGuiKey.PageDown] = (int)Key.PageDown;
-            io.KeyMap[(int)ImGuiKey.Home] = (int)Key.Home;
-            io.KeyMap[(int)ImGuiKey.End] = (int)Key.End;
-            io.KeyMap[(int)ImGuiKey.Delete] = (int)Key.Delete;
-            io.KeyMap[(int)ImGuiKey.Backspace] = (int)Key.BackSpace;
-            io.KeyMap[(int)ImGuiKey.Enter] = (int)Key.Enter;
-            io.KeyMap[(int)ImGuiKey.Escape] = (int)Key.Escape;
-            io.KeyMap[(int)ImGuiKey.Space] = (int)Key.Space;
-            io.KeyMap[(int)ImGuiKey.A] = (int)Key.A;
-            io.KeyMap[(int)ImGuiKey.C] = (int)Key.C;
-            io.KeyMap[(int)ImGuiKey.V] = (int)Key.V;
-            io.KeyMap[(int)ImGuiKey.X] = (int)Key.X;
-            io.KeyMap[(int)ImGuiKey.Y] = (int)Key.Y;
-            io.KeyMap[(int)ImGuiKey.Z] = (int)Key.Z;
-            io.KeyMap[(int)ImGuiKey.Space] = (int)Key.Space;
         }
 
         private void RenderImDrawData(ImDrawDataPtr draw_data, GraphicsDevice gd, CommandList cl)
@@ -701,10 +437,9 @@ namespace ImGuiNET
                 _indexBuffer = gd.ResourceFactory.CreateBuffer(new BufferDescription((uint)(totalIBSize * 1.5f), BufferUsage.IndexBuffer | BufferUsage.Dynamic));
             }
 
-            Vector2 pos = draw_data.DisplayPos;
             for (int i = 0; i < draw_data.CmdListsCount; i++)
             {
-                ImDrawListPtr cmd_list = draw_data.CmdListsRange[i];
+                ImDrawListPtr cmd_list = draw_data.CmdLists[i];
 
                 cl.UpdateBuffer(
                     _vertexBuffer,
@@ -725,14 +460,14 @@ namespace ImGuiNET
             // Setup orthographic projection matrix into our constant buffer
             ImGuiIOPtr io = ImGui.GetIO();
             Matrix4x4 mvp = Matrix4x4.CreateOrthographicOffCenter(
-                pos.X,
-                pos.X + draw_data.DisplaySize.X,
-                pos.Y + draw_data.DisplaySize.Y,
-                pos.Y,
+                0f,
+                io.DisplaySize.X,
+                io.DisplaySize.Y,
+                0.0f,
                 -1.0f,
                 1.0f);
 
-            cl.UpdateBuffer(_projMatrixBuffer, 0, ref mvp);
+            _gd.UpdateBuffer(_projMatrixBuffer, 0, ref mvp);
 
             cl.SetVertexBuffer(0, _vertexBuffer);
             cl.SetIndexBuffer(_indexBuffer, IndexFormat.UInt16);
@@ -746,13 +481,13 @@ namespace ImGuiNET
             int idx_offset = 0;
             for (int n = 0; n < draw_data.CmdListsCount; n++)
             {
-                ImDrawListPtr cmd_list = draw_data.CmdListsRange[n];
+                ImDrawListPtr cmd_list = draw_data.CmdLists[n];
                 for (int cmd_i = 0; cmd_i < cmd_list.CmdBuffer.Size; cmd_i++)
                 {
                     ImDrawCmdPtr pcmd = cmd_list.CmdBuffer[cmd_i];
                     if (pcmd.UserCallback != IntPtr.Zero)
                     {
-                        throw new NotImplementedException();
+                        //throw new NotImplementedException();
                     }
                     else
                     {
@@ -770,8 +505,8 @@ namespace ImGuiNET
 
                         cl.SetScissorRect(
                             0,
-                            (uint)(pcmd.ClipRect.X - pos.X),
-                            (uint)(pcmd.ClipRect.Y - pos.Y),
+                            (uint)pcmd.ClipRect.X,
+                            (uint)pcmd.ClipRect.Y,
                             (uint)(pcmd.ClipRect.Z - pcmd.ClipRect.X),
                             (uint)(pcmd.ClipRect.W - pcmd.ClipRect.Y));
 
