@@ -35,28 +35,31 @@ namespace ImGuiNET
         ResizeGrip = 30,
         ResizeGripHovered = 31,
         ResizeGripActive = 32,
-        Tab = 33,
-        TabHovered = 34,
-        TabActive = 35,
-        TabUnfocused = 36,
-        TabUnfocusedActive = 37,
-        DockingPreview = 38,
-        DockingEmptyBg = 39,
-        PlotLines = 40,
-        PlotLinesHovered = 41,
-        PlotHistogram = 42,
-        PlotHistogramHovered = 43,
-        TableHeaderBg = 44,
-        TableBorderStrong = 45,
-        TableBorderLight = 46,
-        TableRowBg = 47,
-        TableRowBgAlt = 48,
-        TextSelectedBg = 49,
-        DragDropTarget = 50,
-        NavHighlight = 51,
-        NavWindowingHighlight = 52,
-        NavWindowingDimBg = 53,
-        ModalWindowDimBg = 54,
-        COUNT = 55,
+        TabHovered = 33,
+        Tab = 34,
+        TabSelected = 35,
+        TabSelectedOverline = 36,
+        TabDimmed = 37,
+        TabDimmedSelected = 38,
+        TabDimmedSelectedOverline = 39,
+        DockingPreview = 40,
+        DockingEmptyBg = 41,
+        PlotLines = 42,
+        PlotLinesHovered = 43,
+        PlotHistogram = 44,
+        PlotHistogramHovered = 45,
+        TableHeaderBg = 46,
+        TableBorderStrong = 47,
+        TableBorderLight = 48,
+        TableRowBg = 49,
+        TableRowBgAlt = 50,
+        TextLink = 51,
+        TextSelectedBg = 52,
+        DragDropTarget = 53,
+        NavHighlight = 54,
+        NavWindowingHighlight = 55,
+        NavWindowingDimBg = 56,
+        ModalWindowDimBg = 57,
+        COUNT = 58,
     }
 }
