@@ -10,7 +10,8 @@ namespace ImGuiNET
         NavNoCaptureKeyboard = 8,
         NoMouse = 16,
         NoMouseCursorChange = 32,
-        DockingEnable = 64,
+        NoKeyboard = 64,
+        DockingEnable = 128,
         ViewportsEnable = 1024,
         DpiEnableScaleViewports = 16384,
         DpiEnableScaleFonts = 32768,
