@@ -9,7 +9,9 @@ namespace ImGuiNET
         SourceNoHoldToOpenOthers = 4,
         SourceAllowNullID = 8,
         SourceExtern = 16,
-        SourceAutoExpirePayload = 32,
+        PayloadAutoExpire = 32,
+        PayloadNoCrossContext = 64,
+        PayloadNoCrossProcess = 128,
         AcceptBeforeDelivery = 1024,
         AcceptNoDrawDefaultRect = 2048,
         AcceptNoPreviewTooltip = 4096,
