@@ -11,13 +11,18 @@ namespace ImGuiNET
         DockHierarchy = 16,
         AllowWhenBlockedByPopup = 32,
         AllowWhenBlockedByActiveItem = 128,
-        AllowWhenOverlapped = 256,
-        AllowWhenDisabled = 512,
-        NoNavOverride = 1024,
-        RectOnly = 416,
+        AllowWhenOverlappedByItem = 256,
+        AllowWhenOverlappedByWindow = 512,
+        AllowWhenDisabled = 1024,
+        NoNavOverride = 2048,
+        AllowWhenOverlapped = 768,
+        RectOnly = 928,
         RootAndChildWindows = 3,
-        DelayNormal = 2048,
-        DelayShort = 4096,
-        NoSharedDelay = 8192,
+        ForTooltip = 4096,
+        Stationary = 8192,
+        DelayNone = 16384,
+        DelayShort = 32768,
+        DelayNormal = 65536,
+        NoSharedDelay = 131072,
     }
 }
