@@ -6,24 +6,26 @@ namespace ImGuiNET
         None = 0,
         CharsDecimal = 1,
         CharsHexadecimal = 2,
-        CharsUppercase = 4,
-        CharsNoBlank = 8,
-        AutoSelectAll = 16,
-        EnterReturnsTrue = 32,
-        CallbackCompletion = 64,
-        CallbackHistory = 128,
-        CallbackAlways = 256,
-        CallbackCharFilter = 512,
-        AllowTabInput = 1024,
-        CtrlEnterForNewLine = 2048,
-        NoHorizontalScroll = 4096,
-        AlwaysOverwrite = 8192,
-        ReadOnly = 16384,
-        Password = 32768,
+        CharsScientific = 4,
+        CharsUppercase = 8,
+        CharsNoBlank = 16,
+        AllowTabInput = 32,
+        EnterReturnsTrue = 64,
+        EscapeClearsAll = 128,
+        CtrlEnterForNewLine = 256,
+        ReadOnly = 512,
+        Password = 1024,
+        AlwaysOverwrite = 2048,
+        AutoSelectAll = 4096,
+        ParseEmptyRefVal = 8192,
+        DisplayEmptyRefVal = 16384,
+        NoHorizontalScroll = 32768,
         NoUndoRedo = 65536,
-        CharsScientific = 131072,
-        CallbackResize = 262144,
-        CallbackEdit = 524288,
-        EscapeClearsAll = 1048576,
+        CallbackCompletion = 131072,
+        CallbackHistory = 262144,
+        CallbackAlways = 524288,
+        CallbackCharFilter = 1048576,
+        CallbackResize = 2097152,
+        CallbackEdit = 4194304,
     }
 }
