@@ -75,86 +75,99 @@ namespace ImGuiNET
         F10 = 581,
         F11 = 582,
         F12 = 583,
-        Apostrophe = 584,
-        Comma = 585,
-        Minus = 586,
-        Period = 587,
-        Slash = 588,
-        Semicolon = 589,
-        Equal = 590,
-        LeftBracket = 591,
-        Backslash = 592,
-        RightBracket = 593,
-        GraveAccent = 594,
-        CapsLock = 595,
-        ScrollLock = 596,
-        NumLock = 597,
-        PrintScreen = 598,
-        Pause = 599,
-        Keypad0 = 600,
-        Keypad1 = 601,
-        Keypad2 = 602,
-        Keypad3 = 603,
-        Keypad4 = 604,
-        Keypad5 = 605,
-        Keypad6 = 606,
-        Keypad7 = 607,
-        Keypad8 = 608,
-        Keypad9 = 609,
-        KeypadDecimal = 610,
-        KeypadDivide = 611,
-        KeypadMultiply = 612,
-        KeypadSubtract = 613,
-        KeypadAdd = 614,
-        KeypadEnter = 615,
-        KeypadEqual = 616,
-        GamepadStart = 617,
-        GamepadBack = 618,
-        GamepadFaceLeft = 619,
-        GamepadFaceRight = 620,
-        GamepadFaceUp = 621,
-        GamepadFaceDown = 622,
-        GamepadDpadLeft = 623,
-        GamepadDpadRight = 624,
-        GamepadDpadUp = 625,
-        GamepadDpadDown = 626,
-        GamepadL1 = 627,
-        GamepadR1 = 628,
-        GamepadL2 = 629,
-        GamepadR2 = 630,
-        GamepadL3 = 631,
-        GamepadR3 = 632,
-        GamepadLStickLeft = 633,
-        GamepadLStickRight = 634,
-        GamepadLStickUp = 635,
-        GamepadLStickDown = 636,
-        GamepadRStickLeft = 637,
-        GamepadRStickRight = 638,
-        GamepadRStickUp = 639,
-        GamepadRStickDown = 640,
-        MouseLeft = 641,
-        MouseRight = 642,
-        MouseMiddle = 643,
-        MouseX1 = 644,
-        MouseX2 = 645,
-        MouseWheelX = 646,
-        MouseWheelY = 647,
-        ReservedForModCtrl = 648,
-        ReservedForModShift = 649,
-        ReservedForModAlt = 650,
-        ReservedForModSuper = 651,
-        COUNT = 652,
-        ImGuiMod_None = 0,
-        ImGuiMod_Ctrl = 4096,
-        ImGuiMod_Shift = 8192,
-        ImGuiMod_Alt = 16384,
-        ImGuiMod_Super = 32768,
-        ImGuiMod_Shortcut = 2048,
-        ImGuiMod_Mask = 63488,
+        F13 = 584,
+        F14 = 585,
+        F15 = 586,
+        F16 = 587,
+        F17 = 588,
+        F18 = 589,
+        F19 = 590,
+        F20 = 591,
+        F21 = 592,
+        F22 = 593,
+        F23 = 594,
+        F24 = 595,
+        Apostrophe = 596,
+        Comma = 597,
+        Minus = 598,
+        Period = 599,
+        Slash = 600,
+        Semicolon = 601,
+        Equal = 602,
+        LeftBracket = 603,
+        Backslash = 604,
+        RightBracket = 605,
+        GraveAccent = 606,
+        CapsLock = 607,
+        ScrollLock = 608,
+        NumLock = 609,
+        PrintScreen = 610,
+        Pause = 611,
+        Keypad0 = 612,
+        Keypad1 = 613,
+        Keypad2 = 614,
+        Keypad3 = 615,
+        Keypad4 = 616,
+        Keypad5 = 617,
+        Keypad6 = 618,
+        Keypad7 = 619,
+        Keypad8 = 620,
+        Keypad9 = 621,
+        KeypadDecimal = 622,
+        KeypadDivide = 623,
+        KeypadMultiply = 624,
+        KeypadSubtract = 625,
+        KeypadAdd = 626,
+        KeypadEnter = 627,
+        KeypadEqual = 628,
+        AppBack = 629,
+        AppForward = 630,
+        GamepadStart = 631,
+        GamepadBack = 632,
+        GamepadFaceLeft = 633,
+        GamepadFaceRight = 634,
+        GamepadFaceUp = 635,
+        GamepadFaceDown = 636,
+        GamepadDpadLeft = 637,
+        GamepadDpadRight = 638,
+        GamepadDpadUp = 639,
+        GamepadDpadDown = 640,
+        GamepadL1 = 641,
+        GamepadR1 = 642,
+        GamepadL2 = 643,
+        GamepadR2 = 644,
+        GamepadL3 = 645,
+        GamepadR3 = 646,
+        GamepadLStickLeft = 647,
+        GamepadLStickRight = 648,
+        GamepadLStickUp = 649,
+        GamepadLStickDown = 650,
+        GamepadRStickLeft = 651,
+        GamepadRStickRight = 652,
+        GamepadRStickUp = 653,
+        GamepadRStickDown = 654,
+        MouseLeft = 655,
+        MouseRight = 656,
+        MouseMiddle = 657,
+        MouseX1 = 658,
+        MouseX2 = 659,
+        MouseWheelX = 660,
+        MouseWheelY = 661,
+        ReservedForModCtrl = 662,
+        ReservedForModShift = 663,
+        ReservedForModAlt = 664,
+        ReservedForModSuper = 665,
+        COUNT = 666,
+        ModNone = 0,
+        ModCtrl = 4096,
+        ModShift = 8192,
+        ModAlt = 16384,
+        ModSuper = 32768,
+        ModMask = 61440,
         NamedKey_BEGIN = 512,
-        NamedKey_END = 652,
-        NamedKey_COUNT = 140,
-        KeysData_SIZE = 652,
-        KeysData_OFFSET = 0,
+        NamedKey_END = 666,
+        NamedKey_COUNT = 154,
+        KeysData_SIZE = 154,
+        KeysData_OFFSET = 512,
     }
 }
