@@ -9,10 +9,11 @@ namespace ImGuiNET
         MouseButtonMiddle = 2,
         MouseButtonMask = 31,
         MouseButtonDefault = 1,
-        NoOpenOverExistingPopup = 32,
-        NoOpenOverItems = 64,
-        AnyPopupId = 128,
-        AnyPopupLevel = 256,
-        AnyPopup = 384,
+        NoReopen = 32,
+        NoOpenOverExistingPopup = 128,
+        NoOpenOverItems = 256,
+        AnyPopupId = 1024,
+        AnyPopupLevel = 2048,
+        AnyPopup = 3072,
     }
 }
