@@ -4,10 +4,11 @@ namespace ImGuiNET
     public enum ImGuiSelectableFlags
     {
         None = 0,
-        DontClosePopups = 1,
+        NoAutoClosePopups = 1,
         SpanAllColumns = 2,
         AllowDoubleClick = 4,
         Disabled = 8,
-        AllowItemOverlap = 16,
+        AllowOverlap = 16,
+        Highlight = 32,
     }
 }
