@@ -25,11 +25,17 @@ namespace ImGuiNET
         GrabMinSize = 20,
         GrabRounding = 21,
         TabRounding = 22,
-        ButtonTextAlign = 23,
-        SelectableTextAlign = 24,
-        SeparatorTextBorderSize = 25,
-        SeparatorTextAlign = 26,
-        SeparatorTextPadding = 27,
-        COUNT = 28,
+        TabBorderSize = 23,
+        TabBarBorderSize = 24,
+        TabBarOverlineSize = 25,
+        TableAngledHeadersAngle = 26,
+        TableAngledHeadersTextAlign = 27,
+        ButtonTextAlign = 28,
+        SelectableTextAlign = 29,
+        SeparatorTextBorderSize = 30,
+        SeparatorTextAlign = 31,
+        SeparatorTextPadding = 32,
+        DockingSeparatorSize = 33,
+        COUNT = 34,
     }
 }
