@@ -6,7 +6,7 @@ namespace ImGuiNET
         None = 0,
         Selected = 1,
         Framed = 2,
-        AllowItemOverlap = 4,
+        AllowOverlap = 4,
         NoTreePushOnOpen = 8,
         NoAutoOpenOnLog = 16,
         DefaultOpen = 32,
@@ -17,7 +17,9 @@ namespace ImGuiNET
         FramePadding = 1024,
         SpanAvailWidth = 2048,
         SpanFullWidth = 4096,
-        NavLeftJumpsBackHere = 8192,
+        SpanTextWidth = 8192,
+        SpanAllColumns = 16384,
+        NavLeftJumpsBackHere = 32768,
         CollapsingHeader = 26,
     }
 }
