@@ -20,15 +20,13 @@ namespace ImGuiNET
         NoBringToFrontOnFocus = 8192,
         AlwaysVerticalScrollbar = 16384,
         AlwaysHorizontalScrollbar = 32768,
-        AlwaysUseWindowPadding = 65536,
-        NoNavInputs = 262144,
-        NoNavFocus = 524288,
-        UnsavedDocument = 1048576,
-        NoDocking = 2097152,
-        NoNav = 786432,
+        NoNavInputs = 65536,
+        NoNavFocus = 131072,
+        UnsavedDocument = 262144,
+        NoDocking = 524288,
+        NoNav = 196608,
         NoDecoration = 43,
-        NoInputs = 786944,
-        NavFlattened = 8388608,
+        NoInputs = 197120,
         ChildWindow = 16777216,
         Tooltip = 33554432,
         Popup = 67108864,
